changed headers, added constructors, cleaned some code

+++ src/Update.cpp
@@ -1,39 +1,35 @@
-//
-// Created by donat on 5/30/2021.
-//
-
 #include "Engine.h"
 
 void Engine::update() {
-    //Update snake position
-    if(timeSinceLastMove.asSeconds() >= seconds(1.f / float(speed)).asSeconds()){
+    //Update Snake position
+    if (timeSinceLastMove.asSeconds() >= seconds(1.f / float(speed)).asSeconds()) {
         Vector2f thisSectionPosition = snake[0].getPosition();
         Vector2f lastSectionPosition = thisSectionPosition;
 
         //
-        if(!directionQueue.empty()){
-            //Make sure the snake doesn't go into apposite direction
-            switch(snakeDirection){
+        if (!directionQueue.empty()) {
+            //Make sure the Snake doesn't go into apposite direction
+            switch (snakeDirection) {
                 case Direction::UP:
-                    if (directionQueue.front() != Direction::DOWN){
+                    if (directionQueue.front() != Direction::DOWN) {
                         snakeDirection = directionQueue.front();
 
                     }
                     break;
                 case Direction::DOWN:
-                    if (directionQueue.front() != Direction::UP){
+                    if (directionQueue.front() != Direction::UP) {
                         snakeDirection = directionQueue.front();
 
                     }
                     break;
                 case Direction::LEFT:
-                    if (directionQueue.front() != Direction::RIGHT){
+                    if (directionQueue.front() != Direction::RIGHT) {
                         snakeDirection = directionQueue.front();
 
                     }
                     break;
                 case Direction::RIGHT:
-                    if (directionQueue.front() != Direction::LEFT){
+                    if (directionQueue.front() != Direction::LEFT) {
                         snakeDirection = directionQueue.front();
 
                     }
@@ -46,16 +42,16 @@ void Engine::update() {
         score += snake.size() + (applesEatenTotal + 1);
         scoreText.setString(to_string(score));
         FloatRect scoreTextBounds = scoreText.getLocalBounds();
-        scoreText.setPosition(resolution.x - scoreTextBounds.width -15, -9);
+        scoreText.setPosition(resolution.x - scoreTextBounds.width - 15, -9);
 
-        //Do we need to grow the snake
-        if(sectionsToAdd){
+        //Do we need to grow the Snake
+        if (sectionsToAdd) {
             addSnake();
             sectionsToAdd--;
         }
 
         //Update snakes head position
-        switch (snakeDirection){
+        switch (snakeDirection) {
             case Direction::RIGHT:
                 snake[0].setPosition(Vector2f(thisSectionPosition.x + 20, thisSectionPosition.y));
                 break;
@@ -70,53 +66,53 @@ void Engine::update() {
                 break;
         }
 
-        //Update snake tail position
-        for (int s=1; s< snake.size(); s++){
+        //Update Snake tail position
+        for (size_t s = 1; s < snake.size(); s++) {
             thisSectionPosition = snake[s].getPosition();
             snake[s].setPosition(lastSectionPosition);
             lastSectionPosition = thisSectionPosition;
         }
 
-        // update snake functions
-        for(auto & s : snake){
+        // update Snake functions
+        for (auto &s: snake) {
             s.update();
         }
 
         //Collision detection with Apple
-        if (snake[0].getShape().getGlobalBounds().intersects(apple.getSprite().getGlobalBounds())){
-            applesEatenThisLevel +=1;
-            applesEatenTotal +=1;
+        if (snake[0].getShape().getGlobalBounds().intersects(apple.getSprite().getGlobalBounds())) {
+            applesEatenThisLevel += 1;
+            applesEatenTotal += 1;
             applesEatenText.setString("apples " + to_string(applesEatenTotal));
             FloatRect currentLevelTextBounds = currentLevelText.getGlobalBounds();
-            applesEatenText.setPosition(Vector2f(currentLevelTextBounds.left + currentLevelTextBounds.width +20, -9));
+            applesEatenText.setPosition(Vector2f(currentLevelTextBounds.left + currentLevelTextBounds.width + 20, -9));
 
             //see if need to load new level
             bool beginningNewLevel = false;
             if (applesEatenThisLevel >= 10) {
                 //Begin the next level if there are more levels, else just play current level
-                if(currentLevel < maxLevels) {
+                if (currentLevel < maxLevels) {
                     beginningNewLevel = true;
                     beginNextLevel();
                 }
             }
             if (!beginningNewLevel) {
-                //We hit the apple, increase speed/snake size
-                sectionsToAdd +=4;
+                //We hit the Apple, increase speed/Snake size
+                sectionsToAdd += 4;
                 speed++;
                 moveApple();
             }
         }
 
-        //collision with snake body
-        for( int s = 1; s < snake.size(); s++){
-            if (snake[0].getShape().getGlobalBounds().intersects(snake[s].getShape().getGlobalBounds())){
+        //collision with Snake body
+        for (size_t s = 1; s < snake.size(); s++) {
+            if (snake[0].getShape().getGlobalBounds().intersects(snake[s].getShape().getGlobalBounds())) {
                 //game over
                 currentGameState = GameState::GAMEOVER;
             }
         }
 
-        //Collision with wall
-        for (auto & w : wallSections) {
+        //Collision with Wall
+        for (auto &w: wallSections) {
             if (snake[0].getShape().getGlobalBounds().intersects(w.getShape().getGlobalBounds())) {
                 //Game over
                 currentGameState = GameState::GAMEOVER;
